move proxy into core (as part of internal redirection), leaving the configuration logic at the handler-level
