+++ lib/core/config.c
@@ -89,6 +89,7 @@ void h2o_config_init(h2o_globalconf_t *config)
     config->http1.upgrade_to_http2 = H2O_DEFAULT_HTTP1_UPGRADE_TO_HTTP2;
     config->http1.callbacks = H2O_HTTP1_CALLBACKS;
     config->http2.idle_timeout = H2O_DEFAULT_HTTP2_IDLE_TIMEOUT;
+    config->proxy.io_timeout = H2O_DEFAULT_PROXY_IO_TIMEOUT;
     config->http2.max_concurrent_requests_per_connection = H2O_HTTP2_SETTINGS_HOST.max_concurrent_streams;
     config->http2.max_streams_for_priority = 16;
     config->http2.callbacks = H2O_HTTP2_CALLBACKS;
